add test-openai

+++ t2i23-test-stable.py
@@ -35,9 +35,9 @@ t2i_pipe = DiffusionPipeline.from_pretrained(
     dtype=torch.float16
 ).to("cuda")
 # t2i_pipe = StableDiffusion3Pipeline.from_pretrained("stabilityai/stable-diffusion-3.5-medium", dtype=torch.float16).to("cuda")
-# t2i_pipe.transformer = t2i_pipe.transformer.half()
-# t2i_pipe.vae = t2i_pipe.vae.half()
-# t2i_pipe.text_encoder = t2i_pipe.text_encoder.half()
+t2i_pipe.unet = t2i_pipe.unet.half()
+t2i_pipe.vae = t2i_pipe.vae.half()
+t2i_pipe.text_encoder = t2i_pipe.text_encoder.half()
 i23_pipeline = TrellisImageTo3DPipeline.from_pretrained("microsoft/TRELLIS-image-large")
 i23_pipeline.cuda()
